Allow arbitrary URIs for clients; fix compile for older Rust.
Rather than enforcing a two-level "apps, then clients" hierarchy, just
make the configuration have a single list of URIs. This gives the
adminstratrator complete flexibility for the URIs and simplifies the
code. There is no parsing of the path, there is no longer a nested
HashMap. Everything is smaller and simpler.

Also, add an "extern crate" for serde_yaml and prefix it with ::
to prevent a compile error seen on Rust 1.29.1 on DragonFlyBSD.